+++ spec/unit/scenarios/saving-records.spec.ts
@@ -1,7 +1,7 @@
 import { specRequest as request } from '@rvoh/psychic-spec-helpers'
+import { PsychicApp } from '../../../src/index.js'
 import PsychicServer from '../../../src/server/index.js'
 import User from '../../../test-app/src/app/models/User.js'
-import { PsychicApp } from '../../../src/index.js'
 
 describe('a visitor attempts to save a record', () => {
   beforeEach(async () => {
@@ -22,9 +22,9 @@ describe('a visitor attempts to save a record', () => {
   })
 
   context('with a record that is invalid at Dream validation level', () => {
-    it('does not save, returns 422', async () => {
-      const res = await request.post('/failed-to-save-test', 422)
-      expect(res.body).toEqual({ type: 'validator', errors: { email: ['contains'] } })
+    it('does not save, returns 400', async () => {
+      const res = await request.post('/failed-to-save-test', 400)
+      expect(res.body).toEqual({})
     })
   })
 
@@ -56,10 +56,7 @@ describe('a visitor attempts to save a record', () => {
         const response = await request.post('/users', 400, {
           data: { user: { email: 123, password: 'howyadoin', name: 456 } },
         })
-        expect(response.body).toEqual({
-          type: 'validator',
-          errors: { email: ['expected string'], name: ['expected string'] },
-        })
+        expect(response.body).toEqual({})
         expect(await User.count()).toEqual(0)
       })
     },