+++ src/openapi-renderer/defaults.ts
@@ -86,21 +86,7 @@ export const DEFAULT_OPENAPI_COMPONENT_RESPONSES = {
   // 400
   BadRequest: {
     description:
-      'The server would not process the request due to something the server considered to be a client error, such as a model validation failure or an openapi validation failure',
-    content: {
-      'application/json': {
-        schema: {
-          oneOf: [
-            {
-              $ref: '#/components/schemas/ValidationErrors',
-            },
-            {
-              $ref: '#/components/schemas/OpenapiValidationErrors',
-            },
-          ],
-        },
-      },
-    },
+      'The server would not process the request due to something the server considered to be a client error',
   },
 
   // 401